# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_yamovo_contentx_internal_services.UpdateArticleRequest:
    properties:
      allow_comment:
        type: boolean
      canonical_url:
        type: string
      category_id:
        type: integer
      content:
        type: string
      excerpt:
        type: string
      featured_image:
        type: string
      format:
        type: string
      is_featured:
        type: boolean
      is_pinned:
        type: boolean
      meta_desc:
        type: string
      meta_keywords:
        type: string
      meta_title:
        type: string
      og_image:
        type: string
      password:
        type: string
      post_type:
        type: string
      published_at:
        type: string
      revision_note:
        type: string
      robots_follow:
        type: boolean
      robots_index:
        type: boolean
      scheduled_at:
        type: string
      slug:
        type: string
      status:
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      template:
        type: string
      title:
        type: string
      visibility:
        type: string
    type: object
...
  internal_handlers.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      err_code:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/internal_handlers.Meta'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: High-performance Headless CMS - API-first content platform
  title: ContentX API
  version: "1.0"
paths:
...
  /articles:
    get:
      description: Returns a paginated, filtered list of articles
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Filter by status
        enum:
        - draft
        - published
        - pending
        - scheduled
        - trash
        - archived
        in: query
        name: status
        type: string
      - description: Post type
        enum:
        - post
        - page
        in: query
        name: post_type
        type: string
      - description: Filter by category ID
        in: query
        name: category_id
        type: string
      - description: Filter by tag slug
        in: query
        name: tag
        type: string
      - description: Search keyword
        in: query
        name: search
        type: string
      - default: newest
        description: Sort order
        enum:
        - newest
        - oldest
        - most_viewed
        - most_liked
        in: query
        name: sort
        type: string
      - description: Filter by author ID
        in: query
        name: author_id
        type: string
      - description: 'Include full article body (content). Default false: list omits
          content for smaller payloads.'
        in: query
        name: full
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List articles
      tags:
      - Articles
    post:
      consumes:
      - application/json
      description: Create a new article (requires articles.create permission)
      parameters:
      - description: Article data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Article'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Create article
      tags:
      - Articles
  /articles/{id}:
    delete:
      description: Soft-delete an article (requires articles.delete permission)
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete article
      tags:
      - Articles
    get:
      description: Returns a single article by its ID
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Article'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get article by ID
      tags:
      - Articles
    put:
      consumes:
      - application/json
      description: Update an existing article (requires articles.edit permission)
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_yamovo_contentx_internal_services.UpdateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Article'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update article
      tags:
      - Articles
  /articles/{id}/approve:
    post:
      description: Approves a pending article, transitioning it to published. Requires
        editor role.
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Article'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Approve article
      tags:
      - Articles
...
  /articles/{id}/publish:
    post:
      description: Publishes an article (status transition to published). Requires
        articles.edit permission.
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Article'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Publish article
      tags:
      - Articles
  /articles/{id}/revisions:
    get:
      description: Returns the revision history for an article
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List article revisions
      tags:
      - Articles
  /articles/{id}/revisions/{revision_id}/restore:
    post:
      description: Restores an article to a specific revision
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Revision ID
        in: path
        name: revision_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Restore article revision
      tags:
      - Articles
  /articles/{id}/schedule:
    post:
      consumes:
      - application/json
      description: Marks an article for automatic publication at the given time. Requires
        articles.edit permission.
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Schedule payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Article'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Schedule article publication
      tags:
      - Articles
...
  /articles/slug/{slug}:
    get:
      description: Returns a single article by its URL slug (public)
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Article'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      summary: Get article by slug
      tags:
      - Articles
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"